Move user resolver to its own package
