+++ pkg/security/resolvers/selinux/selinux_resolver.go
@@ -6,7 +6,7 @@
 //go:build linux
 // +build linux
 
-package resolvers
+package selinux
 
 import (
 	"os/exec"
