+++ pkg/security/resolvers/user/user_resolver.go
@@ -6,7 +6,7 @@
 //go:build linux
 // +build linux
 
-package resolvers
+package user
 
 import (
 	"os/user"
